renamed compactionlease to compaction, parameterized add-license-headers makefile target
