+++ controllers/compaction/add.go
@@ -1,4 +1,18 @@
-package compactionlease
+// Copyright 2023 SAP SE or an SAP affiliate company
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package compaction
 
 import (
 	druidv1alpha1 "github.com/gardener/etcd-druid/api/v1alpha1"
@@ -10,10 +24,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
 
-const compactionLeaseControllerName = "compaction-lease-controller"
+const compactionControllerName = "compaction-controller"
 
 func (r *Reconciler) AddToManager(mgr ctrl.Manager) error {
-	c, err := controller.New(compactionLeaseControllerName, mgr, controller.Options{
+	c, err := controller.New(compactionControllerName, mgr, controller.Options{
 		Reconciler:              r,
 		MaxConcurrentReconciles: r.config.Workers,
 	})
